pkg/loop: consolidate files; improve Mounter interface
