+++ pkg/mount/mount.go
@@ -7,6 +7,11 @@ package mount
 
 // Mounter is an object that can be mounted.
 type Mounter interface {
-	Mount() error
+	// Mount mounts the file system at path.
+	Mount(path string, flags uintptr) error
+
+	// Unmount unmounts a file system that was previously mounted.
+	//
+	// Mount must have been previously called on this same object.
 	Unmount(flags int) error
 }
